web_aupa/requirements.txt
streamlit
requests
psycopg2-binary
python-dotenv